+++ Gwathra/src/Utility/OBJMesh.h
@@ -4,15 +4,17 @@
 #include "bvhtree.h"
 
 namespace gwa {
+	// TODO Implement submeshes
 	class ObjMesh {
 	public:
 		std::vector<float> vertices = std::vector<float>();
 		std::vector<float> uvs = std::vector<float>();
 		std::vector<float> normals = std::vector<float>();
 
-		std::vector<size_t> vertexIndices = std::vector<size_t>();
-		std::vector<size_t> normalIndices = std::vector<size_t>();
-		std::vector<size_t> uvIndices = std::vector<size_t>();
+		std::vector<uint32_t> vertexIndices = std::vector<uint32_t>();
+		std::vector<uint32_t> normalIndices = std::vector<uint32_t>();
+		std::vector<uint32_t> uvIndices = std::vector<uint32_t>();
+	
 
 
 		void setIndOffset(int indexOffset) { m_indexOffset = indexOffset; }
